// src/app/models/vehicle-drivers/vehicle.model.ts
// export interface Vehicle {
//   id: number;
//   make: string;
//   model: string;
//   year: number;
//   licensePlate: string;
//   status: 'Available' | 'On Trek' | 'Under Maintenance';
//   currentDriverId: string; //staff Id (driver)
//   breakdownStatus: boolean;
//   createdAt: Date;
// }

export interface Vehicle {
  _id?: string;            // Optional because it will be generated by the backend
  vehicleNumber: string;   // Unique identifier for the vehicle
  status: 'Available' | 'On Trek' | 'Under Maintenance';  // Status of the vehicle
  returnDate?: string | Date; 
  updatedAt?: string | Date;
}